requirements.txt
alembic==1.4.2
amqp==2.6.1
attrs==20.1.0
Authlib==0.14.3
billiard==3.6.3.0
blis==0.4.1
cachetools==4.1.1
catalogue==1.0.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.1
cycler==0.10.0
cymem==2.0.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
Flask==1.1.2
Flask-HTTPAuth==4.1.0
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
flower==0.9.5
google-api-core==1.22.1
google-auth==1.21.0
google-cloud-core==1.4.1
google-cloud-logging==1.15.1
google-cloud-pubsub==1.7.0
google-cloud-secret-manager==1.0.0
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.31.0
humanize==2.6.0
idna==2.10
ijson==3.1.1
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
kombu==4.6.11
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.1
mockito==1.2.1
more-itertools==8.5.0
murmurhash==1.0.2
numpy==1.19.1
packaging==20.4
pandas==1.1.1
Pillow==6.1.0
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.8.0
protobuf==3.13.0
psycopg2-binary==2.8.5
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-file==1.5.1
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.10.1
simple-salesforce==1.10.1
six==1.15.0
spacy==2.3.2
SQLAlchemy==1.3.19
srsly==1.0.2
thinc==7.4.1
threadpoolctl==2.1.0
tldextract==2.2.3
toml==0.10.1
tornado==6.0.4
tqdm==4.48.2
urllib3==1.25.10
vine==1.3.0
wasabi==0.8.0
Werkzeug==1.0.1
zipp==3.1.0

# Better environment variables 
envparse==0.2.0
